+++ lib/property.ts
@@ -1,13 +1,10 @@
 import { Field, FieldOptions, ReturnTypeFunc } from '@nestjs/graphql';
-import { Prop, PropOptions, SchemaFactory } from '@nestjs/mongoose';
-import { CreateInputType, CreateInputTypeWithin, OutputType, UpdateInputType } from './type-functions';
-import { ValidateIf, ValidateNested } from 'class-validator';
-import { Schema } from 'mongoose';
-import { Type } from 'class-transformer';
+import { Prop, PropOptions } from '@nestjs/mongoose';
 
-import { GraphQLObjectId } from './shared';
+import { FilterOperator, GraphQLObjectId } from './shared';
 import * as util from './util';
 import { MetaKey, Metadata } from './metadata';
+import { Thunk } from './thunk';
 
 type OverrideOptions = Partial<FieldOptions> & { type?: ReturnTypeFunc };
 
@@ -73,246 +70,6 @@ export function Id() {
     Property({ type: () => GraphQLObjectId, create: Skip, db: Skip })(target, propertyKey);
   };
 }
-
-type ThunkScope = 'all' | 'create' | 'update' | 'input' | 'output';
-
-type ThunkOptions = {
-  scopes: Array<ThunkScope> | ThunkScope;
-};
-
-export const hasScope = (option: ThunkOptions, checkScope: ThunkScope) => {
-  const mapping = {
-    all: ['create', 'update', 'output'],
-    input: ['create', 'update'],
-  };
-
-  const normalizedScopes = util.isArray(option.scopes) ? option.scopes : [option.scopes];
-
-  for (const scope of normalizedScopes) {
-    if (mapping[scope as string]?.includes(checkScope)) return true;
-    if (scope === checkScope) return true;
-  }
-
-  return false;
-};
-
-export function Thunk(
-  fn: any,
-  options: ThunkOptions = { scopes: 'all' },
-): PropertyDecorator & MethodDecorator {
-  return (target: object, propertyKey: string | symbol) => {
-    const property = Metadata.for(target).with(propertyKey);
-    const prevThunks = property.get(MetaKey.Thunk);
-    const newThunks = util.defaultTo(prevThunks, []).concat({ fn, options });
-    property.set(MetaKey.Thunk, newThunks);
-  };
-}
-
-export type EmbeddedConfig = {
-  typeFunction: () => any;
-  options: {
-    allowApis: Array<'findAll' | 'findOne' | 'create' | 'update' | 'remove'>;
-    onSubSchema?: (subSchema: Schema) => void;
-    resolverDecorators?: {
-      default?: MethodDecorator | MethodDecorator[];
-      write?: MethodDecorator | MethodDecorator[];
-      read?: MethodDecorator | MethodDecorator[];
-      findOne?: MethodDecorator | MethodDecorator[];
-      findAll?: MethodDecorator | MethodDecorator[];
-      remove?: MethodDecorator | MethodDecorator[];
-      update?: MethodDecorator | MethodDecorator[];
-      create?: MethodDecorator | MethodDecorator[];
-    };
-  };
-};
-
-export function Embedded(typeFunction: EmbeddedConfig['typeFunction'], options?: EmbeddedConfig['options']) {
-  return (target: object, propertyKey: string | symbol) => {
-    const subSchema = SchemaFactory.createForClass(typeFunction());
-    subSchema.virtual('id').get(function () {
-      return (this['_id'] as any).toHexString();
-    });
-    options?.onSubSchema?.(subSchema);
-
-    const isArray = Reflect.getMetadata(MetaKey.DesignType, target, propertyKey) === Array;
-    Prop({ type: isArray ? [subSchema] : subSchema })(target, propertyKey);
-    Thunk(
-      Field(() => (isArray ? [OutputType(typeFunction())] : OutputType(typeFunction())), { nullable: true }),
-      { scopes: 'output' },
-    )(target, propertyKey);
-    Thunk(
-      Field(() => (isArray ? [CreateInputType(typeFunction())] : CreateInputType(typeFunction())), {
-        nullable: true,
-      }),
-      { scopes: 'create' },
-    )(target, propertyKey);
-    Thunk(
-      Field(() => (isArray ? [UpdateInputType(typeFunction())] : UpdateInputType(typeFunction())), {
-        nullable: true,
-      }),
-      { scopes: 'update' },
-    )(target, propertyKey);
-    if (isArray) {
-      Thunk(ValidateNested({ each: true }), { scopes: 'input' })(target, propertyKey);
-    } else {
-      Thunk(ValidateNested(), { scopes: 'input' })(target, propertyKey);
-      Thunk(ValidateIf((_, value) => value !== null))(target, propertyKey);
-    }
-    Thunk(
-      Type(() => UpdateInputType(typeFunction())),
-      { scopes: 'update' },
-    )(target, propertyKey);
-    Thunk(
-      Type(() => CreateInputType(typeFunction())),
-      { scopes: 'create' },
-    )(target, propertyKey);
-
-    Metadata.for(target)
-      .with(propertyKey)
-      .set<EmbeddedConfig>(MetaKey.EmbeddedType, {
-        typeFunction,
-        options: util.defaultTo(options, {
-          allowApis: ['findAll', 'findOne', 'create', 'update', 'remove'],
-        }),
-      });
-  };
-}
-
-export type ReferencesManyConfig = {
-  typeFunction: () => any;
-  options: {
-    from: string;
-    allowCreateWithin?: boolean;
-    noPopulation?: boolean;
-  };
-};
-
-export function ReferencesMany(
-  typeFunction: ReferencesManyConfig['typeFunction'],
-  options: ReferencesManyConfig['options'],
-) {
-  return (target: object, propertyKey: string | symbol) => {
-    Metadata.for(target)
-      .with(propertyKey)
-      .set<ReferencesManyConfig>(MetaKey.ReferencesManyType, { typeFunction, options });
-
-    if (options.allowCreateWithin) {
-      Thunk(
-        Field(() => [CreateInputType(typeFunction())], { nullable: true }),
-        { scopes: 'create' },
-      )(target, propertyKey);
-
-      Thunk(
-        Type(() => CreateInputType(typeFunction())),
-        { scopes: 'create' },
-      )(target, propertyKey);
-    }
-  };
-}
-
-export type HasOneConfig = {
-  typeFunction: () => any;
-  options: {
-    to: string;
-    allowCreateWithin?: boolean;
-    noPopulation?: boolean;
-  };
-};
-export function HasOne(typeFunction: HasOneConfig['typeFunction'], options: HasOneConfig['options']) {
-  return (target: object, propertyKey: string | symbol) => {
-    Metadata.for(target).with(propertyKey).set<HasOneConfig>(MetaKey.HasOneType, { typeFunction, options });
-    if (options.allowCreateWithin) {
-      Thunk(
-        Field(() => CreateInputTypeWithin(typeFunction(), target.constructor, options.to), {
-          nullable: true,
-        }),
-        { scopes: 'create' },
-      )(target, propertyKey);
-    }
-
-    if (options.noPopulation !== true) {
-      Thunk(
-        Field(() => OutputType(typeFunction()), { nullable: true }),
-        { scopes: 'output' },
-      )(target, propertyKey);
-    }
-  };
-}
-
-export type HasManyConfig = {
-  typeFunction: () => any;
-  options: {
-    to: string;
-    allowCreateWithin?: boolean;
-    allowPaginate?: boolean;
-    allowFindAll?: boolean;
-  };
-};
-export function HasMany(typeFunction: HasManyConfig['typeFunction'], options: HasManyConfig['options']) {
-  return (target: object, propertyKey: string | symbol) => {
-    Metadata.for(target).with(propertyKey).set<HasManyConfig>(MetaKey.HasManyType, { typeFunction, options });
-    if (options.allowCreateWithin) {
-      Thunk(
-        Field(() => [CreateInputTypeWithin(typeFunction(), target.constructor, options.to)], {
-          nullable: true,
-        }),
-        { scopes: 'create' },
-      )(target, propertyKey);
-    }
-  };
-}
-
-export type BelongsToConfig = {
-  typeFunction: () => any;
-  options: {
-    from: string;
-    noPopulation?: boolean;
-    strict?: boolean;
-  };
-};
-export function BelongsTo(
-  typeFunction: BelongsToConfig['typeFunction'],
-  options: BelongsToConfig['options'],
-) {
-  return (target: object, propertyKey: string | symbol) => {
-    Metadata.for(target)
-      .with(propertyKey)
-      .set<BelongsToConfig>(MetaKey.BelongsToType, { typeFunction, options });
-  };
-}
-
-export type FilterOperator =
-  | 'eq'
-  | 'in'
-  | 'notEq'
-  | 'notIn'
-  | 'contains'
-  | 'notContains'
-  | 'gt'
-  | 'gte'
-  | 'lt'
-  | 'lte'
-  | 'regex'
-  | 'notRegex'
-  | 'all'
-  | 'exists';
-
-export const allOperators: FilterOperator[] = [
-  'eq',
-  'in',
-  'notEq',
-  'notIn',
-  'contains',
-  'notContains',
-  'gt',
-  'gte',
-  'lt',
-  'lte',
-  'regex',
-  'notRegex',
-  'all',
-  'exists',
-];
 
 export function Filterable(typeFn: () => any, input: { operators: FilterOperator[] }) {
   return (target: object, propertyKey: string | symbol) => {
