+++ lib/relations/index.ts
@@ -1,0 +1,5 @@
+export * from './belongs-to';
+export * from './embedded';
+export * from './has-many';
+export * from './has-one';
+export * from './references-many';
